+++ src/pages/posting/components/NewPost/NewPost.style.ts
