+++ src/pages/posting/components/NewPost/NewPostConfirm.tsx
