+++ src/pages/posting/components/NewPost/index.ts
@@ -1,0 +1,3 @@
+import NewPost from './NewPost';
+
+export { NewPost };
