+++ src/pages/posting/components/PassPosting/PassPosting.style.ts
