+++ src/pages/posting/components/PassPosting/PassPostingConfirm.tsx
