+++ src/pages/posting/components/PassPosting/index.ts
@@ -1,0 +1,3 @@
+import PassPosting from './PassPosting';
+
+export { PassPosting };
