--- src/pages/posting/components/index.ts
@@ -1,6 +1,0 @@
-import NewPost from './NewPost';
-import PassPosting from './PassPosting';
-import NewPostConfirm from './NewPostConfirm';
-import PassPostingConfirm from './PassPostingConfirm';
-
-export { NewPost, PassPosting, NewPostConfirm, PassPostingConfirm };
